Added bytestring InputStream support
This passes all existing tests, and finally allows us to pass in UTF-8
characters (previously this was limited by the InputStream takeByte, which
chokes badly on characters that didn't fit in one byte). This allows us to
test that XID_START and XID_CONTINUE are working properly in the lexer.

This commit is not without reservations:

  * extracting a UTF8 Char involves converting ByteString to Text
  * and to make that efficient, we need to use lazy Text/ByteString, which
    is probably less performant
  * and adds `text` as a dependency