`lexRust` new semantics
`lexRust` now accumulates all whitespace tokens leading to a non-whitepsace token and packages these
with the latter. This will allow us to lex `<<` as two `<` tokens while differentiating it in the
parsing from `< <` by checking if the second `<` token has and whitespace preceding it. See this
question I asked for more details:

<http://stackoverflow.com/questions/40538124/pushing-back-tokens-in-happy-and-alex>

Accordingly, some changes where made to the representation of whitespace at the token level. The
tests where also updated.

The actual changes involving `<<` and `<` are not in this commit.