Correct handling of literals
Literals are now correctly parsed and represented as their Haskell counterparts in the AST. This
involved figuring out the escape sequences in strings, formats, etc. The AST and parser were adjusted
accordingly. All of the work for the conversion occurs in the new `Syntax.Constants` file.

One thought for the future: is representing floats using `Double` a good idea? I'm not sure yet.
This may result in parsing and pretty printing not being inverse operations!