Adding tests for parsing
Added 'ParserTest' module with some starting tests for parsing types and patterns, most of which are
currently failing. This involved tweaking the AST again. This should make debugging the parser much
easier (ie. not just looking at the output of --info on Happy). It looks like paths are currently a
pain point for both patterns and types.