Pretty print literals
This commit includes a couple of small changes in the parser that individually weren't worth a commit. The pretty printer now deals with literals properly, and a test suite for these has been set up. On the AST side, `FloatTy` was also bunched into the `Suffix` type (where `IntTy` and `UIntTy` already were).