Finished and tested printing of literals and types
Roughly everthing "sort of" works at this point, which makes it much easier to make in depth
progress on anything. This commit introduces pretty-printing and testing of literals and types.
There might still be some work to do on tweaking 'printType' to work well for very large types that
need to be broken onto multiple lines.