Polished Pretty
With pretty-printing pretty much completely working, I moved the bulk of the functions into an
internal module, then from 'Language.Rust.Pretty' itself, I only expose the typeclasses 'Pretty' and
'PrettyAnnotated', and a bunch of instances. Ideally, this is going to be enough for most users -
the functionality is exposed, and the Haddock will be short.

Also internally cleaned up some stray functions in Pretty (block-like things are now named
accordingly and mostly unified).