Literals and Attributes
In the latest attempt to make a working parser, I am going to slowly migrate the contents of Parser2
into Parser. Whie I do this, I will

  * make sure everything is thoroughly tested
  * make sure there are no conflicts in the grammar

Also, I am taking the approach similar to language-c for annotations. Applicative syntax is cool but
painful, ugly (esp. when escaping every $), and inefficient (I only need the hi and lo). To this
end, the `Located` class of things that have a `Span` has been added.

This commit adds literals and attributes to Parser.y