Parse paths/returns/ifs for expressions
Solved the conflict issue around adding (qualified and unqualified) path expressions and return
expressions (with an actual expression following them). The problem was actually in `stmt` in
relation to 'Semi' and 'NoSemi'.

Also added support for expressions not requiring semicolons, namely 'if', 'if let', 'while',
'while let', 'loop', 'for' and 'match'. Also added support for macro expressions.

All in all, introduced only one S/R conflict around returns with or without expressions. The
conflict appears to be doing the right thing, however. The total number of conflicts is 3 S/R
conflicts.