Keep as many multi-character operators as possible
A while ago, I switched to tokenizing '<<' as two '<' tokens (with the second one recgonizing there
is no space in front of it). Not only that, but I went (almost) all out and converted all multi-character tokens into single tokens. For example, '-=' is '-' then '='. Turns out this is quite annoying now. The new plan is:

  * Change as many of these back to what they were. Keep off the problematic tokens which started
    the whole '<<' vs '<' '<' change. Those are '>>=', '||', '&&', '>=', '<<', and '>>'.
  * Make the parser pass information back to the lexer about what tokens are expected at a given
    state.

This commit takes care of the first (way simpler) point. The number of conflicts halves, as does the
number of failing tests. Tests for the lexer will also have to be fixed (but I'll worry about that
more when I get to the second point).