Updated AST of types
Renamed 'ObjectSum' since it is just 'TraitObject'. Also removed 'ImplicitSelf' - since the recent
work on paths, this is just a special case of a path. Also updated equality predicates in where
clauses to support arbitrary types on the LHS and RHS.

In the parser, found several instances where we were using 'ty' instead of 'ty_sum'.