Fix types for parsing a full file
Fixed the ugly hack of shoe-horning the result of a parse into a 'crate'. Removed the Crate (it
makes no sense in a file by file situation) and MacroDef definitions and added a SourceFile one.

Added support for parsing, pretty printing, and resolving SourceFile's and updated examples
accordingly. This also now handles shebangs properly (see #14).