Wrap up 'Spans'
This commit adds some testing of 'Span' (at least on the handful of core types for which we are
also testing parsing). The testing procedure is:

  1. parse the initial values
  2. go through all of its subfields recursively
  3. for each one, use the parsed span to extract a substring ffrom the initial input
  4. re-parse that substring and make sure it matches

While doing this, I found and fixed a bunch of small edge cases. There are probably a couple tiny
ones still lurking about (and one with a TODO), but I now feel comfortable saying this
fixes #14 and addresses some points in #16.