First pass at revamping 'Pretty'
Note that this is not quite complete - there may be some cases where we want more newlines than we
currently have. That said, this fixes some smaller issues and unifies abstractions (see 'block' in
particular).

Also, 'wl-pprint-annotated' is currently broken around 'annotate'. I've opened an [issue][0] and a
[PR][1] to fix this. In the meantime, 'annotate' has been redefined to be a no-op.

The end goal of this branch is to bring #11 to completion.

[0]: https://github.com/minad/wl-pprint-annotated/issues/2
[1]: https://github.com/minad/wl-pprint-annotated/pull/6