Where clauses Part II and other multi line stuff
- Where clauses are now fixed on structs and enums. In the process of
fixing these, I stumbled across a parser error too. In the process of
fixing this error, one of the 3 remaining S/R conflicts disappeared.

- Other short multi-line changes:

  * types have also been mostly fixed.
  * function return types are never on a line of their own
  * method calls don't use 'hang' (even if the 'self' ends up as more
    than one line, those lines shouldn't be indented like the method
    call)
  * patterns in match statements can break along '|'
  * there are trailing commas in matches (and on blocks)

- Method calls and fields break in the same way (and chain together),
and continue to chain even with indexing '[]', trys '?', and tuple
fields '.0' are sprinkled in.

- Single arguments in function/method calls can be multiline without
breaking up that call. For example, this is now valid output:

      foo(bar(baz(boo(far(faz(
        looooooooooooooong,
        arrrrrrrrrrrrrrrgs,
      ))))))

This closes #11. That said, we should continue to improve
pretty-printing to match [fmt-rfcs][0].

  [0]: https://github.com/rust-lang-nursery/fmt-rfcs