Support new type paths (`Vec::<i32>`)
Previously, only expressions could (and had to) put a `::`
discriminator between identifiers and generics (so as to disambiguate
with the less than operator). Now, type paths can do this too (although
they do not _have_ to).

The parsing paths for type and expression paths are now much similar.

Also fixed a bounds issue on trait aliases (`trait Foo = ?Send` is now
allowed).