Add support for unquoting within quotation blocks.
This (sadly large) patch extends Language.Rust.Quote with support for
unquotation. Unquotation allows the programmer to write a large quoted
block, but substitute in items from the surrounding Haskell context.
This is really handy when constructing expressions or statements based
on programmer input.

At the moment, support is limited to a handful of unquotation
situations: expressions, statements, and sets of statements within a
block. In addition, unquotation support is currently limited to
identifiers; a nice addition would be to allow simple expressions (like
calls to functions) within an unquoting section, like Template Haskell
allows.